requirements.txt
aniso8601==9.0.1
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
Faker==18.4.0
flasgger==0.9.5
Flask==2.2.3
Flask-RESTful==0.3.9
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mistune==2.0.5
packaging==23.1
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
psycopg==3.1.8
psycopg2==2.9.6
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-simplexml==0.1.5
pytz==2023.3
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.9
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
